+++ tests/cli/fmg/test_cli_fmg_main.py
@@ -5,7 +5,7 @@ Testing the cli app
 from typer.testing import CliRunner
 
 from fotoobo.cli.main import app
-from fotoobo.helpers.cli import parse_help_output
+from tests.helper import parse_help_output
 
 runner = CliRunner()