+++ CHANGELOG.md
@@ -7,6 +7,13 @@ For examples and guidelines see [https://keepachangelog.com/](https://keepachang
 ## [Unreleased]
 
 
+### Changed
+
+- Switch from easysnmp to pysnmp because pysnmp is pure Python
+- Split the cli command `fgt` `confcheck` to subcommand: `fgt` `config` `check` as there will be
+  more subcommands in the future
+
+
 ## [Released]
 
 ## [0.5.0] - 2023-03-02
@@ -14,7 +21,6 @@ For examples and guidelines see [https://keepachangelog.com/](https://keepachang
 ### Added
 
 - fotoobo is now on the official PyPI
-- 
 
 ## [0.4.3-6] - 2023-03-02